+++ assets/javascripts/kitten/components/form/text-input-with-button/index.js
@@ -166,14 +166,13 @@ export const TextInputWithButton = React.forwardRef(
             {
               'k-Form-TextInputWithButton__button--valid': valid,
               'k-Form-TextInputWithButton__button--error': error,
-              'k-Form-TextInputWithButton__button--disabled': disabled,
+              'k-Form-TextInputWithButton__button--disabled': disabled || buttonProps.disabled,
             },
           )}
-          disabled={disabled}
+          disabled={disabled || buttonProps.disabled}
           variant={variant}
           modifier={modifier}
           size={size}
-          disabled={disabled}
         >
           {value || buttonValue}
         </ButtonComponent>
@@ -202,7 +201,9 @@ TextInputWithButton.defaultProps = {
   disabled: false,
   size: 'regular',
   buttonValue: 'Button',
-  buttonProps: {},
+  buttonProps: {
+    disabled: false,
+  },
   modifier: 'beryllium',
   rounded: false,
   inset: false,