+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 - Fix: `TextInput`: Improve placeholder style.
 - Feature: `ColorSelect`: Add component.
+- Fix: `TextInputWithButton`: Fix `disabled` prop priority.
 
 ## [6.17.0] - 2022-01-10